AP-17775: Implement proof-of-concept configure and execute of Python Nodes
To implement configure, we needed to convert the DataTableSpec into an
AnnotatedColumnarValueSchema (new in org.knime.core.table)
which we then serialize into JSON and send it to Python.

For execute() we have copied a lot of methods from the Python3ScriptingKernelBackend
because we need putTable/getTable methods etc. We should refactor this code to prevent
duplication. The tables are made available on the Python side the same way as in a script node.

AP-17775 (Deep Dive: Node API)